Re-factor cmd functions to library
