+++ cmd/car/root.go
@@ -2,27 +2,18 @@ package main
 
 import (
 	"fmt"
-	"os"
 
-	carv2 "github.com/ipld/go-car/v2"
+	"github.com/ipld/go-car/cmd/car/lib"
 	"github.com/urfave/cli/v2"
 )
 
 // CarRoot prints the root CID in a car
 func CarRoot(c *cli.Context) (err error) {
-	inStream := os.Stdin
-	if c.Args().Len() >= 1 {
-		inStream, err = os.Open(c.Args().First())
-		if err != nil {
-			return err
-		}
-	}
-
-	rd, err := carv2.NewBlockReader(inStream)
+	roots, err := lib.CarRoot(c.Args().First())
 	if err != nil {
 		return err
 	}
-	for _, r := range rd.Roots {
+	for _, r := range roots {
 		fmt.Printf("%s\n", r.String())
 	}
